Fix VerticalTicker resize
