+++ lib/VerticalTicker.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useMemo, useRef } from "react";
 import { animate } from "./animate";
 import { TickerProps } from "./TickerProps";
+import { useElementSize } from "./useElementSize";
 
 export const VerticalTicker: React.FC<TickerProps> = ({
   children,
@@ -21,9 +22,9 @@ export const VerticalTicker: React.FC<TickerProps> = ({
     [duration, easing, delay]
   );
 
-  useEffect(() => {
-    const trackHeight = track1.current?.getBoundingClientRect().height;
+  const { height: trackHeight } = useElementSize(track1);
 
+  useEffect(() => {
     if (!trackHeight || !track1.current || !track2.current) {
       return;
     }
@@ -73,7 +74,7 @@ export const VerticalTicker: React.FC<TickerProps> = ({
     return () => {
       controller.abort();
     };
-  }, [options]);
+  }, [trackHeight, options]);
 
   return (
     <div
